+++ src/cli/actions/index.ts
@@ -1,3 +1,4 @@
 export * from "./milestone-3.js";
 export * from "./shared.js";
 export * from "./milestone-4.js";
+export * from "./milestone-9.js";